+++ rust/dbn/src/record_buf.rs
@@ -168,6 +168,13 @@ impl<const CAP: usize> RecordBuf<CAP> {
         }
     }
 
+    /// Returns a reference to the common record header at the start of every record.
+    pub fn header(&self) -> &RecordHeader {
+        // SAFETY: `RecordBuf` always holds a valid record. The `hd` field of the union
+        // is always valid because every record starts with a `RecordHeader`.
+        unsafe { &self.0.hd }
+    }
+
     /// Returns `true` if the buffer holds a record of type `T`.
     pub fn has<T: HasRType>(&self) -> bool {
         T::has_rtype(self.header().rtype)
@@ -335,10 +342,32 @@ impl<const CAP: usize> RecordBuf<CAP> {
 }
 
 impl<const CAP: usize> Record for RecordBuf<CAP> {
-    fn header(&self) -> &RecordHeader {
-        // SAFETY: `RecordBuf` always holds a valid record. The `hd` field of the union
-        // is always valid because every record starts with a `RecordHeader`.
-        unsafe { &self.0.hd }
+    fn record_size(&self) -> usize {
+        self.header().record_size()
+    }
+
+    fn rtype(&self) -> crate::Result<RType> {
+        self.header().rtype()
+    }
+
+    fn raw_rtype(&self) -> u16 {
+        self.header().rtype as u16
+    }
+
+    fn publisher_id(&self) -> u16 {
+        self.header().publisher_id
+    }
+
+    fn publisher(&self) -> crate::Result<crate::Publisher> {
+        self.header().publisher()
+    }
+
+    fn instrument_id(&self) -> u64 {
+        self.header().instrument_id as u64
+    }
+
+    fn raw_ts_event(&self) -> u64 {
+        self.header().ts_event
     }
 
     fn raw_index_ts(&self) -> u64 {